final commit

+++ frontend/workout/src/components/WorkoutDetails.jsx
@@ -1,30 +1,37 @@
-import React from 'react'
-import { useWorkoutsContext } from '../hooks/useWorkoutContext'
-// date fns
-import formatDistanceToNow from 'date-fns/formatDistanceToNow'
+import { useWorkoutsContext } from "../hooks/useWorkoutContext"
+import { useAuthContext } from "../hooks/useAuthContext"
+
 const WorkoutDetails = ({ workout }) => {
-    const { dispatch } = useWorkoutsContext()
-    const handleClick = async () => {
-        const response = await fetch(`http://localhost:4000/api/workouts/${workout._id}`, {
-            method: 'DELETE'
-        })
-        const json = await response.json()
-      
-        if (response.ok) {
-            dispatch({type: 'DELETE_WORKOUT', payload: json})
-        }
-     console.log(json)
+  const { dispatch } = useWorkoutsContext()
+  const { user } = useAuthContext()
+
+  const handleClick = async () => {
+    if (!user) {
+      return
     }
+
+    const response = await fetch(`http://localhost:4000/api/workouts/${workout._id}`, {
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
+    })
+
+    const json = await response.json()
+
+    if (response.ok) {
+      dispatch({ type: 'DELETE_WORKOUT', payload: json })
+    }
+  }
+
   return (
-    <div className='workout-details'>
+    <div className="workout-details">
       <h4>{workout.title}</h4>
-      <p><strong>Load (kg):</strong> {workout.load}</p>
-      <p><strong>Reps: </strong> {workout.reps}</p>
-      {/* <p>{new Date(workout.createdAt).toLocaleString()}</p> */}
-      <p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
-      <span className="material-symbols-outlined" onClick={handleClick}>delete</span>
+      <p><strong>Load (kg): </strong>{workout.load}</p>
+      <p><strong>Reps: </strong>{workout.reps}</p>
+      <p>{workout.createdAt}</p>
+      <span onClick={handleClick}>🗑️</span>
     </div>
   )
 }
-
 export default WorkoutDetails
